+++ Editor/GameCode/NewScriptDialog.xaml.cs
@@ -40,6 +40,7 @@ namespace Editor.GameCode
             if (!Validate()) return;
 
             IsEnabled = false;
+            loadingSpinner.Visibility = Visibility.Visible;
 
             try
             {
@@ -117,6 +118,8 @@ namespace Editor.GameCode
 
         private static void CreateScript(string name, string path, string solution, string projectName)
         {
+            //Thread.Sleep(300000000);
+
             if (!Directory.Exists(path)) Directory.CreateDirectory(path);
 
             var cpp = Path.GetFullPath(Path.Combine(path, $"{name}.cpp"));