+++ Changelog.md
@@ -2,7 +2,10 @@
 
 ## UI
  - Scalar and Vector controls
- - Upgraded visuals for Transform Component, Engine Path Dialog
+ - Upgraded visuals for Transform Component, Engine Path Dialog, New Script Dialog
+
+## Editor
+ - Custom .NET App host application to fix engine initialization error
 
 # v0.3.3-beta - 2025.6.10.